Face Generator and Bank Account
This project introduces the concepts of classes by having the user manipulate a face that is drawn using Turtle. Its characteristics can be manipulated based on the user and can be reset due to object orientation.
The Bank Account also uses object orientation as the account can have a constantly manipulated number.