Class Chessboard Generator, Password Checker
This Project uses the same logic previously used on the function based chessboard in which the user can choose the size of the chessboard and the internal squares adapt to formal the new size. This time it is done using object orientation by using classes.
A password checker was also created to give familiarity with checking user inputs.