Operator Overloading, Random Visualization, and Custom Shapes
This project explores how to overload operators in order to compare sizes of user given cuboids. It also gives visualization of how the build in random for Python is not truly random as there can be a predictable outcome. The Custom Shapes project allows the user to add and remove shapes of their choice on a canvas (i.e. Turtle), giving more practice with object oriented programming.